# backend/user-service/docs/swagger.yaml
basePath: /
definitions:
  models.UserResponse:
    properties:
      age:
        type: integer
      avatar_url:
        type: string
      bio:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_organizer:
        type: boolean
      last_name:
        type: string
      username:
        type: string
    type: object
  services.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  services.RegisterInput:
    properties:
      age:
        minimum: 18
        type: integer
      email:
        type: string
      first_name:
        type: string
      is_organizer:
        type: boolean
      last_name:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - age
    - email
    - first_name
    - last_name
    - password
    - username
    type: object
  services.SendFriendRequestInput:
    properties:
      receiver_id:
        type: integer
    required:
    - receiver_id
    type: object
  services.UpdateProfileInput:
    properties:
      bio:
        type: string
      first_name:
        type: string
      last_name:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: API for Meetup social networking application
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Meetup App API
  version: "1.0"
paths:
  /friends:
    get:
      description: Get all friends of the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get friends list
      tags:
      - Friends
  /friends/{id}:
    delete:
      description: Remove a friend from the current user's friend list
      parameters:
      - description: Friend ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove friend
      tags:
      - Friends
  /friends/request:
    post:
      consumes:
      - application/json
      description: Send a friend request to another user
      parameters:
      - description: Friend request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.SendFriendRequestInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Send friend request
      tags:
      - Friends
  /friends/request/{id}/accept:
    put:
      description: Accept a pending friend request
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Accept friend request
      tags:
      - Friends
  /friends/request/{id}/reject:
    put:
      description: Reject a pending friend request
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reject friend request
      tags:
      - Friends
  /login:
    post:
      consumes:
      - application/json
      description: Login with email and password
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/services.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - Authentication
  /profile:
    get:
      description: Get the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update the profile of the currently authenticated user
      parameters:
      - description: Profile data
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/services.UpdateProfileInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
  /profile/avatar:
    put:
      consumes:
      - application/json
      description: Update the avatar URL of the currently authenticated user
      parameters:
      - description: Avatar URL
        in: body
        name: avatar
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
      security:
      - BearerAuth: []
      summary: Update user avatar
      tags:
      - Users
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user with email, username, password and profile
        information
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/services.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      summary: Register new user
      tags:
      - Authentication
  /users/search:
    get:
      description: Search for users by username, first name or last name
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Results limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Search users
      tags:
      - Users
  /users/suggestions:
    get:
      description: Get suggested friends for the current user
      parameters:
      - description: Results limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Get friend suggestions
      tags:
      - Friends
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"